horzvert_timeseries.py: find_pairs.py folded in
